Add UTF-8 validation functions, to check filenames with
Filenames in RAR might be stored as ASCII, UTF-8 or UTF-16 (compressed
in the latter case). We convert them into UTF-8. To make sure the
conversion is correct (and the data was valid to begin with), there's
now two functions to help with validating the results.

dmc_unrar_unicode_is_valid_utf8() checks that a given string contains
valid UTF-8 data and dmc_unrar_unicode_make_valid_utf8() enforces this
by cutting the string off at the first invalid sequence.